Create MySQL for FastAPI
When there are over N notices today, it's kinda tricky to get last index of yeterday, this causes showing no notices when I call "어제 공지".

So, with using MySQL for FastAPI, I managed to solve "어제 공지 보여줘" option.

For now, the db is just for calling yesterday notices but will change it to work with users' last notice id to call the pareser less.

Hosted MySQL server with AWS RDS (free tier)